+++ packages/permissionless/CHANGELOG.md
@@ -1,5 +1,23 @@
 # permissionless
 
+## 0.1.38
+
+### Patch Changes
+
+- 71740c2d9a3fbb289df277831bde10b33995d9df: Fixed batchcall for 7579 accounts
+
+## 0.1.37
+
+### Patch Changes
+
+- 75f48b038f2c0943c1df3e0aa693a641fb4be941: Fixed typo erc7569 > erc7579.
+
+## 0.1.36
+
+### Patch Changes
+
+- f3b5d5e65c34479c8b7a3e59d96fd3fb8bc4933c: Made ExecutionMode fields optional
+
 ## 0.1.35
 
 ### Patch Changes